support large avatar size on ipad (no 'regular' layout)
