+++ Sources/Controllers/Profile/Views/ProfileAvatarView.swift
@@ -6,8 +6,7 @@ public class ProfileAvatarView: ProfileBaseView {
 
     public struct Size {
         static let height: CGFloat = 271
-        static let avatarWidth: CGFloat = 180
-        static let avatarHeight: CGFloat = 180
+        static let avatarSize: CGFloat = 180
         static let whiteBarHeight: CGFloat = 60
     }
 
@@ -53,8 +52,7 @@ extension ProfileAvatarView {
         addSubview(avatarImageView)
 
         avatarImageView.snp_makeConstraints { make in
-            make.width.equalTo(Size.avatarWidth)
-            make.height.equalTo(Size.avatarHeight)
+            make.width.height.equalTo(Size.avatarSize)
             make.centerX.equalTo(self)
             make.bottom.equalTo(self)
         }
@@ -68,7 +66,7 @@ extension ProfileAvatarView {
 
     override public func layoutSubviews() {
         super.layoutSubviews()
-        avatarImageView.layer.cornerRadius = Size.avatarWidth / 2
+        avatarImageView.layer.cornerRadius = Size.avatarSize / 2
     }
 
     public func prepareForReuse() {
